Change over to 2D game & update collision check

+++ core/src/com/game/messages/PushOutMessage.java
@@ -3,11 +3,11 @@ package com.game.messages;
 import com.game.objects.GameObject;
 
 public class PushOutMessage extends GameMessage{
-	public double whereX;
-	public double whereY;
+	public float whereX;
+	public float whereY;
 	
 	
-	public PushOutMessage (GameObject object, double whereX, double whereY){
+	public PushOutMessage (GameObject object, float whereX, float whereY){
 		this.type = MessageType.pushOut;
 		this.object = object;
 		this.whereX = whereX;
